+++ sumologic/sumologic_monitors_library_monitor.go
@@ -146,6 +146,7 @@ type MonitorsLibraryMonitor struct {
 	Status                  []string               `json:"status"`
 	GroupNotifications      bool                   `json:"groupNotifications"`
 	Playbook                string                 `json:"playbook,omitempty"`
+	AutomatedPlaybookIds    []string               `json:"automatedPlaybookIds,omitempty"`
 	AlertName               string                 `json:"alertName,omitempty"`
 	SloID                   string                 `json:"sloId,omitempty"`
 	NotificationGroupFields []string               `json:"notificationGroupFields,omitempty"`